Replace Command abstract class with interface
