+++ src/seedu/addressbook/commands/ClearCommand.java
@@ -1,9 +1,16 @@
 package seedu.addressbook.commands;
 
+import java.util.List;
+
+import seedu.addressbook.data.AddressBook;
+import seedu.addressbook.data.person.ReadOnlyPerson;
+
 /**
  * Clears the address book.
  */
-public class ClearCommand extends Command {
+public class ClearCommand implements Command {
+    protected AddressBook addressBook;
+    protected List<? extends ReadOnlyPerson> relevantPersons;
 
     public static final String COMMAND_WORD = "clear";
     public static final String MESSAGE_USAGE = "Clears address book permanently.\n"
@@ -19,5 +26,13 @@ public class ClearCommand extends Command {
     public CommandResult execute() {
         addressBook.clear();
         return new CommandResult(MESSAGE_SUCCESS);
+    }
+
+    /**
+     * Supplies the data the command will operate on.
+     */
+    public void setData(AddressBook addressBook, List<? extends ReadOnlyPerson> relevantPersons) {
+        this.addressBook = addressBook;
+        this.relevantPersons = relevantPersons;
     }
 }
