+++ src/seedu/addressbook/commands/ExitCommand.java
@@ -1,9 +1,16 @@
 package seedu.addressbook.commands;
 
+import java.util.List;
+
+import seedu.addressbook.data.AddressBook;
+import seedu.addressbook.data.person.ReadOnlyPerson;
+
 /**
  * Terminates the program.
  */
-public class ExitCommand extends Command {
+public class ExitCommand implements Command {
+    protected AddressBook addressBook;
+    protected List<? extends ReadOnlyPerson> relevantPersons;
 
     public static final String COMMAND_WORD = "exit";
 
@@ -16,6 +23,14 @@ public class ExitCommand extends Command {
      */
     public CommandResult execute() {
         return new CommandResult(MESSAGE_EXIT_ACKNOWEDGEMENT);
+    }
+
+    /**
+     * Supplies the data the command will operate on.
+     */
+    public void setData(AddressBook addressBook, List<? extends ReadOnlyPerson> relevantPersons) {
+        this.addressBook = addressBook;
+        this.relevantPersons = relevantPersons;
     }
 
     public static boolean isExit(Command command) {
