+++ src/seedu/addressbook/commands/IncorrectCommand.java
@@ -1,10 +1,16 @@
 package seedu.addressbook.commands;
 
+import java.util.List;
+
+import seedu.addressbook.data.AddressBook;
+import seedu.addressbook.data.person.ReadOnlyPerson;
 
 /**
  * Represents an incorrect command. Upon execution, produces some feedback to the user.
  */
-public class IncorrectCommand extends Command{
+public class IncorrectCommand implements Command{
+    protected AddressBook addressBook;
+    protected List<? extends ReadOnlyPerson> relevantPersons;
 
     public final String feedbackToUser;
 
@@ -17,6 +23,14 @@ public class IncorrectCommand extends Command{
      */
     public CommandResult execute() {
         return new CommandResult(feedbackToUser);
+    }
+
+    /**
+     * Supplies the data the command will operate on.
+     */
+    public void setData(AddressBook addressBook, List<? extends ReadOnlyPerson> relevantPersons) {
+        this.addressBook = addressBook;
+        this.relevantPersons = relevantPersons;
     }
 
 }
